seperate cmd and html
