+++ config/ssg-config.json
